Minimize use of trait objects
The old code was written using trait objects everywhere, which made it
difficult to maintain.  The new code uses concrete enums, which are far
simpler.  Furthermore, the Substrate dependencies have been updated.

Although the new code is longer, it is (in my opinion) easier to modify,
as async/await and the `?` operator can be used as normal.