Use a minimal config for `cargo deny`
This is (almost) the minimum config needed to pass `cargo deny check` on my system. I kept in the `unlicensed` and `copyleft` items even though they are not strictly needed as I felt they communicate something important to the reader. The same might be true for other directives left out here; feedback welcome.